change atomic_flag to use compare_exchange_n under the hood
Using the specialized instruction for TAS only would make sense if we
would assume that compare_exchange for 1 byte data is suboptimal, which
we don't.

From looking at the produced code, compare_exchange_n integrates better
in the surrounding code, since it only checks for the cc flag, and
doesn't need a real transfer of the actual value that is stored in the
flag.